+++ ufo/module/processors/processor.py
@@ -11,14 +11,14 @@ from typing import Dict, List, Optional, Tuple
 from pywinauto.controls.uiawrapper import UIAWrapper
 
 from ufo import utils
-from ufo.agent.agent import AppAgent, HostAgent
-from ufo.agent.memory import MemoryItem
+from ufo.agents.agent.app_agent import AppAgent
+from ufo.agents.agent.host_agent import HostAgent
+from ufo.agents.memory import MemoryItem
 from ufo.automator.ui_control.control_filter import ControlFilterFactory
 from ufo.config.config import Config
 from ufo.module import interactor
 from ufo.module.processors.basic import BaseProcessor
 from ufo.module.state import Status
-
 
 configs = Config.get_instance().config_data
 BACKEND = configs["CONTROL_BACKEND"]
